refactor(ui): move wallet source picker to ui
