+++ simulator/CMakeLists.txt
@@ -186,6 +186,7 @@ set(APP_UI_SOURCES
     ${APP_UI_DIR}/numeric_keypad.c
     ${APP_UI_DIR}/power.c
     ${APP_UI_DIR}/word_selector.c
+    ${APP_UI_DIR}/wallet_source_picker.c
     ${APP_UI_DIR}/battery.c
     ${APP_UI_DIR}/key_info.c
     ${APP_UI_DIR}/sankey.c
@@ -279,7 +280,6 @@ set(APP_PAGE_SOURCES
     ${APP_PAGES_DIR}/shared/key_confirmation.c
     ${APP_PAGES_DIR}/shared/mnemonic_editor.c
     ${APP_PAGES_DIR}/shared/storage_browser.c
-    ${APP_PAGES_DIR}/shared/wallet_source_picker.c
     ${APP_PAGES_DIR}/shared/wipe_flash_dialog.c
     # Top-level pages
     ${APP_PAGES_DIR}/capture_entropy.c
